+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.4
+
+- Set as default service Calibre Web
+- Added Ingress for Calibre Web
+
 ## 0.3
 
 - Set as default service Calibre Server